+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 
 - [User] Log in
 - [User] Create an account
+- [Admin] Add an authorized bot (backend only)